pyproject.toml
[project]
name = "travel-planner-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-adk>=0.2.0",
    "google-generativeai>=0.8.5",
    "mcp-flight-search>=0.2.1",
    "mcp-hotel-search>=0.1.2",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "setuptools>=79.0.0",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
